Support passing ref (React); Fix console error
